fix: изменить метод getReviewRating(long reviewId) в классе ReviewRatingDbStorage т.о., чтобы запрос к БД выполнялся 1 раз

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -18,5 +18,5 @@ public class Review {
     private Long userId;
     @NotNull
     private Long filmId;
-    private int useful;
+    private long useful;
 }
