+++ src/main/java/ru/yandex/practicum/filmorate/storage/dal/ReviewRatingStorage.java
@@ -3,5 +3,5 @@ package ru.yandex.practicum.filmorate.storage.dal;
 public interface ReviewRatingStorage {
     boolean addLikeDislike(long reviewId, long userId, boolean isUseful);
     boolean removeLikeDislike(long reviewId, long userId, boolean isUseful);
-    int getReviewRating(long reviewId);
+    long getReviewRating(long reviewId);
 }
